apps/api-crew-vercel/pyproject.toml
[project]
name = "sa-call-analyzer-crew"
version = "0.1.0"
description = "CrewAI routes for Call Analyzer (separate Vercel deployment)"
requires-python = ">=3.10"
dependencies = [
    # FastAPI core
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    
    # LLM providers
    "anthropic>=0.18.1",
    "openai>=1.12.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.0.5",
    
    # CrewAI
    "crewai>=0.11.0",
    
    # Observability
    "arize-otel>=0.1.0",
    "arize-phoenix-client>=1.0.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "openinference-instrumentation-openai>=0.1.0",
    "openinference-instrumentation-crewai>=0.1.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0",
    "opentelemetry-instrumentation-threading>=0.45b0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "requests>=2.31.0",
    "python-pptx>=0.6.21",
    "python-docx>=1.1.0",
    "rapidfuzz>=3.0.0",
    "google-cloud-bigquery>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.vercel]
entrypoint = "main.py"
